move more checks out of librustc
